Refactors clients into separate classes
